refactor(connection): move Connection impl in respcetive modules
